Ejercicio 3 de la unidad 3 resuelto (corregido).

+++ src/unidad3/Bisiesto.java
@@ -9,7 +9,7 @@ public class Bisiesto {
 	public static void main(String[] args) throws IOException {
 		BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
 		System.out.println("Introduce un año cualquiera");
-		int año = (int) Long.parseLong(in.readLine());
+		int año = Integer.parseInt(in.readLine());
 		
 		if(año%4==0 && (año%100!=0 || año%400==0)) {
 				System.out.print("Es bisiesto");
